pyproject.toml
[tool.poetry]
name = "backend-core"
version = "0.1.0"
description = "Core backend services for personal website"
authors = ["Carlos Crespo <macayaven@gmail.com>"]
readme = "README.md"
packages = [{include = "backend_core"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
pydantic = {extras = ["email"], version = "^2.10.4"}
pydantic-settings = "^2.7.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
invoke = "^2.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.0"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.0"
httpx = "^0.28.1"
autoflake = "^2.3.1"
ruff = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
explicit_package_bases = true
exclude = ".venv|build|dist|.*\\.egg-info"
install_types = true
non_interactive = true
namespace_packages = true

[[tool.mypy.overrides]]
module = [
    "pydantic_settings.*",
    "fastapi.*",
    "jose.*",
    "passlib.*",
    "sqlalchemy.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [".venv", "build", "dist", "*.egg-info"]
unsafe-fixes = true

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E203"]